Updated DriveTowardTargetCommand
I updated the DriveTowardTargetCommand so that it has the structure of a command. I also fixed the logic, commented it for clarity, and put the skeleton of a button binding in the RobotContainer (b/c I'm not sure what button to bind it to yet). I also made it so while the button is held down, the robot tries to locate the AprilTag, drives toward it, then stops driving toward it when it gets to a certain distance, or if the tag is lost. I thought this would be easier/cleaner than "toggling" between standard joystick control and auto-driving.